Bug 1080760 - generate bug suggestions for text_log_summary artifacts
Whenever a ``text_log_summary`` with ``error_lines`` is submitted, generate an
additional ``Bug suggestions`` artifact for it.
