+++ treeherder/webapp/api/artifact.py
@@ -6,6 +6,7 @@ from rest_framework import viewsets
 from rest_framework.response import Response
 from treeherder.webapp.api.utils import UrlQueryFilter, oauth_required
 from treeherder.model.derived import JobsModel, ArtifactsModel
+from treeherder.model.bug_suggestions import get_bug_suggestions_artifacts
 
 
 class ArtifactViewSet(viewsets.ViewSet):
@@ -54,11 +55,20 @@ class ArtifactViewSet(viewsets.ViewSet):
 
     @oauth_required
     def create(self, request, project):
+        artifacts = request.DATA
 
-        job_guids = [x['job_guid'] for x in request.DATA]
-        with JobsModel(project) as jobsModel, ArtifactsModel(project) as artifacts_model:
+        job_guids = [x['job_guid'] for x in artifacts]
+        with JobsModel(project) as jobs_model, ArtifactsModel(project) as artifacts_model:
 
-            job_id_lookup = jobsModel.get_job_ids_by_guid(job_guids)
-            artifacts_model.load_job_artifacts(request.DATA, job_id_lookup)
+            # create an accompanying ``Bug suggestions`` artifact for any
+            # eligible artifacts.
+            tls_list = [x for x in artifacts if x['name'] == 'text_log_summary']
+            bsa = get_bug_suggestions_artifacts(tls_list)
+
+            if bsa:
+                artifacts.extend(bsa)
+
+            job_id_lookup = jobs_model.get_job_ids_by_guid(job_guids)
+            artifacts_model.load_job_artifacts(artifacts, job_id_lookup)
 
             return Response({'message': 'Artifacts stored successfully'})
